// GF2020-v1/src/app/models/account-recievable/order-to-check.model.ts
export class OrderCheckObj {
    ordertocash_settings_pk: number;
    show_on_load: number;
    outstanding_based_on: number;
    zero_os_archive_flag: boolean;
    set_value_archive_flag: boolean;
    set_archive_value: number;
    is_active: number;
    created_by_fk: number;
    created_on: string;
    last_updated_by_fk: number;
    last_updated_on: string;
    version_no: number;
}